Fixing deadlock in `draw_button_cb`
- If we start threads to compute our set and then pause it and then when
drawing if we in sync mode, our draw function will wait forever for
workers to finish. So if this situation happens, draw function checks
`is_paused` variable, shows error about it and returns doing nothing.
- Also now we have new struct binded_widget_pair_t for holding two
widgets (draw function requires `drawing_area` and `error_view` widgets)
- As side effect from adding threads and reallocking workers we got
Undefined Behaviour when draw function tries to wait for workers which
aren't started and not even initialized. To fix this we created
`worker_t` struct which has two fields - thread and is_init flag.
When we start to draw we check which workers are initialized and wait
for these threads. When we start computation process we are flipping
is_init flag to `true`.
- When we are initializing threads we may got undefined behaviour if our
numbers of threads are more than 2**60. Firstly, we can't allocate so much
memory, second if we'll be iterating so much, we will overflow our
workers* pointer and this is Undefined Behaviour.
So we've changed `size_t` num_threads to `int16_t`.